third_party/blink/web_tests/external/wpt/dom/nodes/Document-createEvent.https-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Should throw NOT_SUPPORTED_ERR for pluralized legacy event interface "KeyboardEvents"
  assert_throws_dom: function "function () {\n          var evt = document.createEvent(plural);\n        }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "AnimationEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "CloseEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "ErrorEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "IDBVersionChangeEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "PageTransitionEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "PopStateEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "TrackEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "TransitionEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "WebGLContextEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "WheelEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
Harness: the test ran to completion.
